requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-adminactions==1.6.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
envparse==0.2.0
Pillow==6.1.0
psycopg2-binary==2.8.3
PyJWT==2.1.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
xlrd==2.0.1
xlwt==1.3.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==3.2
multidict==5.1.0
requests==2.26.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3